Moved data.ts into data folder, created resident-data.ts
